requirements.txt
numpy
scipy
pillow
torch
astra-toolbox

